Autoconfigure Java and Maven Publish
These should be separate plugins, but for now we simply throw it in here until we actually have such plugins. This also splits the repository configuration into its own script to make it easier to maintain things. The plan for the repositories is to have them generated via the distribution kit in the future, then it is anyway going to be a separate file.